refactor: 숫자 검증 위치 변경
